requirements.txt
fastapi~=0.112.2
pydantic-settings~=2.6.1
selenium~=4.24.0
dnspython~=2.6.1
aiohttp~=3.11.10
pydantic~=2.10.3
ipwhois~=1.3.0
dnslib~=0.9.26
scapy~=2.6.1
uvicorn~=0.32.1